+++ backend/lib/backend_web/controllers/instance_controller.ex
@@ -7,9 +7,14 @@ defmodule BackendWeb.InstanceController do
   action_fallback(BackendWeb.FallbackController)
 
   def show(conn, %{"id" => domain}) do
-    instance = Api.get_instance!(domain)
-    last_crawl = get_last_crawl(domain)
-    render(conn, "show.json", instance: instance, crawl: last_crawl)
+    instance = Api.get_instance_with_peers(domain)
+
+    if instance == nil or instance.opt_out == true do
+      send_resp(conn, 404, "Not found")
+    else
+      last_crawl = get_last_crawl(domain)
+      render(conn, "show.json", instance: instance, crawl: last_crawl)
+    end
   end
 
   # def update(conn, %{"id" => id, "instance" => instance_params}) do